Added prevent ghost clicks logic, cleaned up other logic.
Prevent Ghost Clicks is functionality to prevent touch events from also triggering a secondary click event along with the primary tap event when a touch ends. My implementation adds this on a per-recognizer basis using the preventGhosts flag on the recognizer options. This flag will need to be set on all recognizer options that you define. I am also considering adding this flag as a manager options to set it for every recognizer associated with that manager. See #17 on GitHub.

I cleaned up the logic in the addRecognizer and setupRecognizerWithOptions to be a bit more fault tolerant. I also added a parameter to setupRecognizerWithOptions so that I can pass the element from the directive's link function and enable prventGhosts from a single location.